Refactor ExpressionParser to multiple files
