+++ src/site/pages/digital/src/containers/ExprToCircuitPopup/index.tsx
@@ -1,36 +1,37 @@
-import React, {useState} from "react";
+import React, { useState } from "react";
 
-import {CreateAddGroupAction} from "core/actions/addition/AddGroupActionFactory";
+import { CreateAddGroupAction } from "core/actions/addition/AddGroupActionFactory";
 
-import {Popup}   from "shared/components/Popup";
+import { Popup }   from "shared/components/Popup";
 
-import {CloseHeaderPopups} from "shared/state/Header";
-import {useSharedDispatch, useSharedSelector} from "shared/utils/hooks/useShared";
+import { CloseHeaderPopups } from "shared/state/Header";
+import { useSharedDispatch, useSharedSelector } from "shared/utils/hooks/useShared";
 
-import {OrganizeMinDepth} from "core/utils/ComponentOrganizers";
+import { OrganizeMinDepth } from "core/utils/ComponentOrganizers";
 
-import {GroupAction}             from "core/actions/GroupAction";
-import {CreateDeselectAllAction,
+import { GroupAction }             from "core/actions/GroupAction";
+import { CreateDeselectAllAction,
         CreateGroupSelectAction,
-        SelectAction,}            from "core/actions/selection/SelectAction";
-import {PlaceAction}             from "core/actions/addition/PlaceAction";
-import {CreateICDataAction}      from "digital/actions/CreateICDataAction";
+        SelectAction, }            from "core/actions/selection/SelectAction";
+import { PlaceAction }             from "core/actions/addition/PlaceAction";
+import { CreateICDataAction }      from "digital/actions/CreateICDataAction";
 
-import {DigitalCircuitInfo}     from "digital/utils/DigitalCircuitInfo";
-import {DigitalComponent}       from "digital/models/DigitalComponent";
-import {ICData}                 from "digital/models/ioobjects/other/ICData";
-import {IC}                     from "digital/models/ioobjects/other/IC";
-import {LED}                    from "digital/models/ioobjects/outputs/LED";
-import {ConstantLow}            from "digital/models/ioobjects/inputs/ConstantLow";
-import {ConstantHigh}           from "digital/models/ioobjects/inputs/ConstantHigh";
-import {Button}                 from "digital/models/ioobjects/inputs/Button";
-import {Switch}                 from "digital/models/ioobjects/inputs/Switch";
-import {Clock}                  from "digital/models/ioobjects/inputs/Clock";
-import {GenerateTokens,
-        ExpressionToCircuit}    from "digital/utils/ExpressionParser";
-import {FormatLabels, FormatMap}    from "digital/utils/ExpressionParser/Constants";
+import { DigitalCircuitInfo }     from "digital/utils/DigitalCircuitInfo";
+import { DigitalComponent }       from "digital/models/DigitalComponent";
+import { ICData }                 from "digital/models/ioobjects/other/ICData";
+import { IC }                     from "digital/models/ioobjects/other/IC";
+import { LED }                    from "digital/models/ioobjects/outputs/LED";
+import { ConstantLow }            from "digital/models/ioobjects/inputs/ConstantLow";
+import { ConstantHigh }           from "digital/models/ioobjects/inputs/ConstantHigh";
+import { Button }                 from "digital/models/ioobjects/inputs/Button";
+import { Switch }                 from "digital/models/ioobjects/inputs/Switch";
+import { Clock }                  from "digital/models/ioobjects/inputs/Clock";
+import { ExpressionToCircuit }    from "digital/utils/ExpressionParser";
+import { GenerateTokens } from "digital/utils/ExpressionParser/GenerateTokens";
+import { FormatLabels, FormatMap }    from "digital/utils/ExpressionParser/Constants";
 
 import "./index.scss";
+
 
 type Props = {
     mainInfo: DigitalCircuitInfo;
